+++ client/src/Commands/registerCommands.ts
@@ -8,6 +8,7 @@ import * as CompileSMCommand from "./compileSM";
 import * as UploadToServerCommand from "./uploadToServer";
 import * as RefreshPluginsCommand from "./refreshPlugins";
 import * as InsertParametersCommand from "./insertParameters";
+import * as setFileAsMainCommand from "./setFileAsMain";
 
 export function registerSMCommands(context: ExtensionContext) {
   let createTask = commands.registerCommand(
@@ -63,4 +64,10 @@ export function registerSMCommands(context: ExtensionContext) {
     InsertParametersCommand.run.bind(undefined)
   );
   context.subscriptions.push(InsertParameters);
+
+  let setFileAsMain = commands.registerTextEditorCommand(
+    "sourcepawn-vscode.setFileAsMain",
+    setFileAsMainCommand.run.bind(undefined)
+  );
+  context.subscriptions.push(setFileAsMain);
 }